health_managment/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
pydantic>=2.0.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0